pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.6",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gradio>=5.12.0",
    "gunicorn>=23.0.0",
    "langchain-openai>=0.3.0",
    "langgraph-checkpoint>=2.0.9",
    "langgraph>=0.2.62",
    "openai>=1.59.6",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "tavily-python>=0.5.0",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.37",
    "pydantic>=2.10.5",
]
